// src/types/DeviceData.d.ts
/**
 * 设备数据类型
 * @interface DeviceData
 * @property {number[]} data1 - 数据1
 * @property {number[]} data2 - 数据2
 * @property {number[]} data3 - 数据3
 * @property {number[]} data4 - 数据4
 */
export interface DeviceData {
  data1: number[]
  data2: number[]
  data3: number[]
  data4: number[]
}